requirements.txt
python-docx==1.1.2
black==24.8.0
pillow==10.4.0
requests
flask
